// src/generated/models/Pulse_apppermissionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pulse_apppermissionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pulse_apppermissionsstatecode = keyof typeof Pulse_apppermissionsstatecode;
export const Pulse_apppermissionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pulse_apppermissionsstatuscode = keyof typeof Pulse_apppermissionsstatuscode;

export interface Pulse_apppermissionsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  "pulse_App@odata.bind"?: string;
  pulse_apppermission1?: string;
  pulse_apppermissionid: string;
  "pulse_SecurityRole@odata.bind"?: string;
  statecode: Pulse_apppermissionsstatecode;
  statuscode?: Pulse_apppermissionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pulse_apppermissions extends Pulse_apppermissionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pulse_appname?: string;
  pulse_securityrolename?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pulse_app?: object;
  _pulse_app_value?: string;
  pulse_securityrole?: object;
  _pulse_securityrole_value?: string;
}