// src/generated/models/RolesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Rolescomponentstate = {
  0: 'Published',
  1: 'Unpublished',
  2: 'Deleted',
  3: 'Deleted Unpublished'
} as const;
export type Rolescomponentstate = keyof typeof Rolescomponentstate;
export const Rolesisautoassigned = {
  0: 'No',
  1: 'Yes'
} as const;
export type Rolesisautoassigned = keyof typeof Rolesisautoassigned;
export const Rolesisinherited = {
  0: 'Team privileges only',
  1: 'Direct User (Basic) access level and Team privileges'
} as const;
export type Rolesisinherited = keyof typeof Rolesisinherited;

export interface RolesBase {
  applicationid?: string;
  appliesto: string;
  "BusinessUnitId@odata.bind": string;
  canbedeleted: string;
  description: string;
  FnOAotName?: string;
  importsequencenumber?: number;
  isautoassigned: Rolesisautoassigned;
  iscustomizable: string;
  isinherited: Rolesisinherited;
  issytemgenerated: boolean;
  name: string;
  overriddencreatedon?: string;
  roleid: string;
  summaryofcoretablepermissions: string;
}

export interface Roles extends RolesBase {
  businessunitidname: string;
  componentstate: Rolescomponentstate;
  createdbyname?: string;
  createdbyyominame?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  isautoassignedname?: string;
  isinheritedname?: string;
  ismanaged: boolean;
  ismanagedname?: string;
  issytemgeneratedname?: string;
  modifiedbyname?: string;
  modifiedbyyominame?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  organizationid: string;
  organizationidname: string;
  overwritetime: string;
  parentroleidname: string;
  parentrootroleidname: string;
  roleidunique: string;
  solutionid: string;
  supportingsolutionid?: string;
  versionnumber?: number;
  businessunitid?: object;
  _businessunitid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  parentroleid?: object;
  _parentroleid_value?: string;
  parentrootroleid?: object;
  _parentrootroleid_value?: string;
  roletemplateid?: object;
  _roletemplateid_value?: string;
}